chore: separate results runners, remove shared runner
